+++ src/main/resources/db/changelog/011__add_user_file_share_expiry_reminder.sql
@@ -1,0 +1,5 @@
+--liquibase formatted sql
+
+--changeset matigfv:11
+ALTER TABLE user_file_shares ADD COLUMN expiry_reminder_sent_at TIMESTAMP NULL;
+--rollback ALTER TABLE user_file_shares DROP COLUMN expiry_reminder_sent_at;